Feat : BattleWindowUI 스크립트 작성 시작

+++ Assets/01.Scrips/4.Entity/1.Player/StatData.cs
@@ -28,6 +28,6 @@ public class StatData : ScriptableObject
 public class StatEntry
 {
     public StatType statType;
-    public float baseValue;
+    public int baseValue;
 }
